app/requirements/base.txt
Django==2
django-cors-headers==2.3.0
django-debug-toolbar==1.9.1
django-environ==0.4.5
django-extensions==2.0.7
django-redis==4.9.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-filter==2.0.0
Fabric3==1.14.post1
flake8==3.5.0
gunicorn==19.8.1
ipython==6.4.0
redis==2.10.6
requests==2.19.1
Werkzeug==0.14.1
celery==4.2.1
psycopg2-binary==2.7.5
django-crispy-forms==1.7.2
django-braces==1.13.0